1.split components and component states, not written in one file to improve composition and easy maintenance; 2 refactor use typescript
